refactored datacommons-tools to datacommons-schema
