+++ datacommons/datacommons-api/src/datacommons/api/responses.py
@@ -6,3 +6,4 @@ class UpdateResponse(BaseModel):
   """
   success: bool
   message: str
+
